requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
pymongo==4.5.0
qdrant-client==1.7.2
neo4j==5.16.0
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.2
pandas==2.1.4
numpy==1.26.0
PyJWT==2.8.0
google-generativeai==0.5.2